Redo packaging using pyproject.toml and src layout (#1)
* Redo package

* Fix version number v0.1.0

* Formatting

* Docstring darglint

* Update imports

* Formatting

---------
