+++ .bumpversion.toml
@@ -1,0 +1,22 @@
+[tool.bumpversion]
+current_version = "v0.1.0"
+commit = true
+commit_args = "--no-verify"
+tag = true
+tag_name = "{new_version}"
+allow_dirty = true
+parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)(\\.(?P<dev>dev)\\d+\\+[-_a-zA-Z0-9]+)?"
+serialize = [
+    "v{major}.{minor}.{patch}.{dev}{distance_to_latest_tag}+{short_branch_name}",
+    "v{major}.{minor}.{patch}"
+]
+message = "Version updated from {current_version} to {new_version}"
+
+[[tool.bumpversion.files]]
+filename = "pyproject.toml"
+
+[[tool.bumpversion.files]]
+filename = "README.md"
+
+[[tool.bumpversion.files]]
+filename = "src/imageruler/__init__.py"
