+++ src/imageruler/__init__.py
@@ -1,0 +1,6 @@
+"""Imageruler."""
+
+__version__ = "v0.1.0"
+
+from imageruler import imageruler as imageruler
+from imageruler import regular_shapes as regular_shapes
